requirements.txt
llama-index
python-dotenv
pandas
pypdf
transformers[torch]
InstructorEmbedding
sentence_transformers